+++ src/main/java/com/mw/cc/service/currencyProviders/CurrencyProvider.java
@@ -1,7 +1,8 @@
 package com.mw.cc.service.currencyProviders;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface CurrencyProvider {
-   Map<String, Double> getCurrencies();
+   Map<String, BigDecimal> getCurrencies();
 }
